+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Data Hub API 36.4.0 (2021-02-08)
+
+
+## API
+
+- **Investment** It is now possible to get a list of large capital investment profiles created in activity-stream using 
+  the following URL `/v3/activity-stream/investment/large-capital-investor-profiles`.
+
+
 # Data Hub API 36.3.0 (2021-01-29)